requirements.txt
pyrogram
tgcrypto
python-dotenv
aiohttp
aiohttp-cors
websockets
